requirements.txt
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
matplotlib>=3.7.0
python-binance>=1.0.19
streamlit>=1.30.0
plotly>=5.18.0
streamlit-autorefresh>=1.0.1
